Create and use a fixture for creating eligibility rows to reduce duplication and improve test readability.
